refactor: extract TagFilter component from TemplateBrowser
Moves tag frequency calculation, sorting, and button rendering into a
self-contained TagFilter component. TemplateBrowser passes filteredTemplates,
activeTags, and the toggle callback as props.
